pkg_info.cygport: correct search order for Perl dependencies
* lib/pkg_info.cygpart: Correct search order for Perl dependencies and
  suppress auto-generation of Perl dependencies when NO_PERL_DEPS is
  defined.

Dependency generation for Perl at least is too simplistic and doesn't
take into account that some modules required or used might actually be
optional.  It tends to generate too long dependency lists that vary
with the Perl distributions already installed.

For starters, the search order should be the reverse of
@INC to skip dependencies that are built-in to perl already, but that
doesn't pick up those modules that are needed with a higher version
since only the presence of the module is detected.  Files in site_perl
shoud never be searched since these are local installs.  Files in
vendor_perl might be useful to check, however due to the version
problem it is better to inject the module dpenedencies from the
cygport file.  So skip those searches when NO_PERL_DEPS is defined,
which it will be for auto-generated cygport files for Perl
distributions (the information is pulled from CPAN/MetaCPAN).