pkg: don't put local user:group info into package archives
Don't leak local user and group into the package archives we make.
(This information is ignored by setup, in any case)

Pick Guest:None to be the owner of all files in the archive, just in
case we ever need to mark some as owned by Administrator:Administrators
in future.