Add some case variants to list of default documentation files installed
Otherwise, if such a file is present with variant case, it is only
installed on case-insensitive systems, making tests behave differently
there.

Add 'Changelog' to the explicit list of docs to package used in
libtxc_dxtn.cygport.

Add documentation files which are now installed into expected filelists

Note that where default_docs contains case variants which are degenerate
on a case-insensitive filesystem, the first one controls the filename of
the installed file on a case-preserving filesystem.

There's no way to always ensure that matches the actual case of the
filename, so filelists must be compared case-insensitively.

Also:
Report filelist differences, rather than just comparision result